Fix Python emitter mapping `numeric` to `int` instead of `float`
Agent-Logs-Url: https://github.com/microsoft/typespec/sessions/5da46ea6-698a-475e-b61f-328680b5ff6a

+++ packages/http-client-python/emitter/src/types.ts
@@ -436,7 +436,7 @@ function emitConstant(context: PythonSdkContext, type: SdkConstantType) {
 }
 
 const sdkScalarKindToPythonKind: Record<string, string> = {
-  numeric: "integer",
+  numeric: "float",
   integer: "integer",
   safeint: "integer",
   int8: "integer",
